split all the rest of `service::auth` into own modules
